+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [1.2.2] - 2016-11-24
 ### Added
 - inspect node & npm versions
 - inspect .gitignore
@@ -27,7 +29,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - initialize this project
 
+[1.2.2]: https://github.com/super-fe/superfe-rn-inspector/compare/1.2.0...1.2.2
 [1.2.0]: https://github.com/super-fe/superfe-rn-inspector/compare/1.1.1...1.2.0
 [1.1.1]: https://github.com/super-fe/superfe-rn-inspector/compare/1.1.0...1.1.1
 [1.1.0]: https://github.com/super-fe/superfe-rn-inspector/compare/1.0.0...1.1.0
-[Unreleased]: https://github.com/super-fe/superfe-rn-inspector/compare/1.2.1...HEAD
+[Unreleased]: https://github.com/super-fe/superfe-rn-inspector/compare/1.2.2...HEAD